requirements.txt
boto3==1.9.224
botocore==1.12.224
Django==2.2.5
django-material-admin==1.4.6
django-storages==1.7.1
docutils==0.15.2
jmespath==0.9.4
Pillow==6.1.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
python-memcached==1.59
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
uWSGI==2.0.18
